+++ crates/affinidi-tdk/common/affinidi-secrets-resolver/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Affinidi Secrets Manager
 
+## xx May 2025 (0.1.8)
+
+* **MAINTENANCE:** Crate dependencies updated
+* **FEATURE:** `get_public_bytes()` and `get_private_bytes()` methods added
+
 ## Release 0.1.7
 
 * Removed drop() trait implementation as was causing the SecretsTask to close early